refactor(learnpage): clean up and modularize code in chinese page
