+++ src/app/(app)/learn/edexcel-igcse/chinese/past-papers/page.tsx
@@ -1,0 +1,11 @@
+"use client";
+
+import PastPapersContent from "../../../../past-paper/page";
+
+const PastPapersPage = () => {  
+  return (
+    <PastPapersContent />
+  );
+};
+
+export default PastPapersPage;
